# Improve performance of sampler experiment