--- a/releasenotes/notes/sampler-performance-81e1649ec4657aad.yaml
+++ b/releasenotes/notes/sampler-performance-81e1649ec4657aad.yaml
@@ -2,4 +2,5 @@
 ---
 upgrade:
   - |
-    Improved performance when the same circuits and multiple parameters are passed to :class:`~.Sampler`.
+    Improved performance when the same circuits and multiple parameters are passed to
+    :class:`~.Sampler`.