--- a/releasenotes/notes/sampler-performance-81e1649ec4657aad.yaml
+++ b/releasenotes/notes/sampler-performance-81e1649ec4657aad.yaml
@@ -1,5 +1,4 @@
 ---
----
 upgrade:
   - |
     Improved performance when the same circuits and multiple parameters are passed to