<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

### Summary

Same idea with https://github.com/Qiskit/qiskit-aer/pull/1840

### Details and comments

Benchmark code:
```python
import numpy as np

from qiskit.circuit.library import EfficientSU2
from qiskit_aer.primitives import Sampler

rng = np.random.default_rng(15)

qc = EfficientSU2(3)
qc.measure_all()

num_circuits = 1000
sampler = Sampler()
parameter_values = rng.random((num_circuits, qc.num_parameters))
circuits = [qc] * num_circuits

%timeit sampler.run(circuits, parameter_values).result()
```
0.12.2: 716 ms ± 1.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
This PR: 534 ms ± 4.33 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)